# Optimize array handling memory usage