--- a/Jint.Benchmark/MinimalScriptBenchmark.cs
+++ b/Jint.Benchmark/MinimalScriptBenchmark.cs
@@ -1,0 +1,7 @@
+﻿namespace Jint.Benchmark
+{
+    public class MinimalScriptBenchmark : SingleScriptBenchmark
+    {
+        protected override string Script => "var done = true;";
+    }
+}