--- a/Jint/Native/Number/Dtoa/CachePowers.cs
+++ b/Jint/Native/Number/Dtoa/CachePowers.cs
@@ -28,8 +28,6 @@
 // Ported to Java from Mozilla's version of V8-dtoa by Hannes Wallnoefer.
 // The original revision was 67d1049b0bf9 from the mozilla-central tree.
 
-using System.Diagnostics;
-
 namespace Jint.Native.Number.Dtoa
 {
 
@@ -52,17 +50,15 @@ namespace Jint.Native.Number.Dtoa
         }
 
 
-        internal static int GetCachedPower(int e, int alpha, int gamma, DiyFp cMk)
+        internal static (short, DiyFp) GetCachedPower(int e, int alpha, int gamma)
         {
             const int kQ = DiyFp.KSignificandSize;
             double k = System.Math.Ceiling((alpha - e + kQ - 1)*Kd1Log210);
             int index = (GrisuCacheOffset + (int) k - 1)/CachedPowersSpacing + 1;
             CachedPower cachedPower = CACHED_POWERS[index];
 
-            cMk.F = cachedPower.Significand;
-            cMk.E = cachedPower.BinaryExponent;
-            Debug.Assert((alpha <= cMk.E + e) && (cMk.E + e <= gamma));
-            return cachedPower.DecimalExponent;
+            //System.Diagnostics.Debug.Assert((alpha <= cMk.E + e) && (cMk.E + e <= gamma));
+            return (cachedPower.DecimalExponent, new DiyFp(cachedPower.Significand, cachedPower.BinaryExponent));
         }
 
         // Code below is converted from GRISU_CACHE_NAME(8) in file "powers-ten.h"