--- a/Jint/Native/Number/Dtoa/CachePowers.cs
+++ b/Jint/Native/Number/Dtoa/CachePowers.cs
@@ -28,9 +28,10 @@
 // Ported to Java from Mozilla's version of V8-dtoa by Hannes Wallnoefer.
 // The original revision was 67d1049b0bf9 from the mozilla-central tree.
 
+using System.Diagnostics;
+
 namespace Jint.Native.Number.Dtoa
 {
-
     public class CachedPowers
     {
         private const double Kd1Log210 = 0.30102999566398114; //  1 / lg(10)
@@ -53,12 +54,13 @@ namespace Jint.Native.Number.Dtoa
         internal static (short, DiyFp) GetCachedPower(int e, int alpha, int gamma)
         {
             const int kQ = DiyFp.KSignificandSize;
-            double k = System.Math.Ceiling((alpha - e + kQ - 1)*Kd1Log210);
-            int index = (GrisuCacheOffset + (int) k - 1)/CachedPowersSpacing + 1;
+            double k = System.Math.Ceiling((alpha - e + kQ - 1) * Kd1Log210);
+            int index = (GrisuCacheOffset + (int) k - 1) / CachedPowersSpacing + 1;
             CachedPower cachedPower = CACHED_POWERS[index];
 
-            //System.Diagnostics.Debug.Assert((alpha <= cMk.E + e) && (cMk.E + e <= gamma));
-            return (cachedPower.DecimalExponent, new DiyFp(cachedPower.Significand, cachedPower.BinaryExponent));
+            var cMk = new DiyFp(cachedPower.Significand, cachedPower.BinaryExponent);
+            Debug.Assert((alpha <= cMk.E + e) && (cMk.E + e <= gamma));
+            return (cachedPower.DecimalExponent, cMk);
         }
 
         // Code below is converted from GRISU_CACHE_NAME(8) in file "powers-ten.h"
@@ -155,7 +157,5 @@ namespace Jint.Native.Number.Dtoa
         };
 
         private const int GrisuCacheOffset = 308;
-
-
     }
 }