… Jint.Benchmark/LinqJsBenchmark.cs
… Jint/Native/Number/Dtoa/DoubleHelper.cs
… Jint/Native/Number/Dtoa/FastDtoa.cs
… Jint/Native/Number/Dtoa/FastDtoaBuilder.cs
… Jint/Runtime/MruPropertyCache2.cs
… Jint/Runtime/TypeConverter.cs