I did first pass on investigating arrays and they memory profile. This is done on top of the benchmark update pull request.

Here are some quick wins from caching common strings, avoiding boxing etc. We are now 40% faster for the famous array test case and use almost 80% less memory (which brings the speed).

*Before*

|   Method | N | ReuseEngine |        Mean |      Error |    StdDev |       Gen 0 |     Gen 1 |     Gen 2 |  Allocated |
|--------- |-- |------------ |------------:|-----------:|----------:|------------:|----------:|----------:|-----------:|
|     **Jint** | **1** |       **False** | **2,081.36 ms** |  **89.654 ms** | **5.0656 ms** | **440500.0000** | **8750.0000** | **2250.0000** | **1784.51 MB** |
|     **Jint** | **1** |        **True** | **2,071.57 ms** | **118.701 ms** | **6.7068 ms** | **441333.3333** | **8916.6667** | **2000.0000** | **1784.37 MB** |

*After*

|   Method | N | ReuseEngine |       Mean |      Error |    StdDev |      Gen 0 |     Gen 1 |     Gen 2 | Allocated |
|--------- |-- |------------ |-----------:|-----------:|----------:|-----------:|----------:|----------:|----------:|
|     **Jint** | **1** |       **False** | **1,201.2 ms** | **137.316 ms** | **7.7586 ms** | **93166.6667** | **4416.6667** | **1250.0000** | **388.35 MB** |
|     **Jint** | **1** |        **True** | **1,187.4 ms** |  **49.602 ms** | **2.8026 ms** | **94000.0000** | **4750.0000** | **1500.0000** | **388.21 MB** |