@lahma 2017-12-19
I've incorporated @ayende 's suggestion, but I will open separate smaller pull requests in hopes of getting tweaks merged.